+++ internal/metrics/reconciler/reconciler.go
@@ -27,14 +27,20 @@ const (
 )
 
 const (
-	skippedOnlyStatusChangeName    = "skipped_only_status_change"
-	existingInProgressName         = "existing_in_progress"
-	failureUnableToGetPodName      = "failure_unable_to_get_pod"
-	failurePodDoesntExistName      = "failure_pod_doesnt_exist"
-	failureConfigurationName       = "failure_configuration"
-	failureValidationName          = "failure_validation"
-	failureStatesDeterminationName = "failure_states_determination"
-	failureStatesActionName        = "failure_states_action"
+	skippedOnlyStatusChangeName = "skipped_only_status_change"
+	existingInProgressName      = "existing_in_progress"
+	failureName                 = "failure"
+)
+
+type FailureReason string
+
+const (
+	FailureReasonUnableToGetPod      = FailureReason("unable_to_get_pod")
+	FailureReasonPodDoesNotExist     = FailureReason("pod_does_not_exist")
+	FailureReasonConfiguration       = FailureReason("configuration")
+	FailureReasonValidation          = FailureReason("validation")
+	FailureReasonStatesDetermination = FailureReason("states_determination")
+	FailureReasonStatesAction        = FailureReason("states_action")
 )
 
 var (
@@ -52,53 +58,17 @@ var (
 		Help:      "Number of attempted reconciles where one was already in progress for the same namespace/name (results in a requeue)",
 	}, []string{})
 
-	failureUnableToGetPod = prometheus.NewCounterVec(prometheus.CounterOpts{
+	failure = prometheus.NewCounterVec(prometheus.CounterOpts{
 		Namespace: metricscommon.Namespace,
 		Subsystem: Subsystem,
-		Name:      failureUnableToGetPodName,
-		Help:      "Number of reconciles where there was a failure to get the pod (results in a requeue)",
-	}, []string{})
-
-	failurePodDoesntExist = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: metricscommon.Namespace,
-		Subsystem: Subsystem,
-		Name:      failurePodDoesntExistName,
-		Help:      "Number of reconciles where the pod was found not to exist (results in failure)",
-	}, []string{})
-
-	failureConfiguration = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: metricscommon.Namespace,
-		Subsystem: Subsystem,
-		Name:      failureConfigurationName,
-		Help:      "Number of reconciles where there was a failure to configure (results in failure)",
-	}, []string{})
-
-	failureValidation = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: metricscommon.Namespace,
-		Subsystem: Subsystem,
-		Name:      failureValidationName,
-		Help:      "Number of reconciles where there was a failure to validate (results in failure)",
-	}, []string{})
-
-	failureStatesDetermination = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: metricscommon.Namespace,
-		Subsystem: Subsystem,
-		Name:      failureStatesDeterminationName,
-		Help:      "Number of reconciles where there was a failure to determine states (results in failure)",
-	}, []string{})
-
-	failureStatesAction = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: metricscommon.Namespace,
-		Subsystem: Subsystem,
-		Name:      failureStatesActionName,
-		Help:      "Number of reconciles where there was a failure to action the determined states (results in failure)",
-	}, []string{})
+		Name:      failureName,
+		Help:      "Number of reconciles where there was a failure",
+	}, []string{metricscommon.ReasonLabelName})
 )
 
 // allMetrics must include all metrics defined above.
 var allMetrics = []prometheus.Collector{
-	skippedOnlyStatusChange, existingInProgress, failureUnableToGetPod, failurePodDoesntExist, failureConfiguration,
-	failureValidation, failureStatesDetermination, failureStatesAction,
+	skippedOnlyStatusChange, existingInProgress, failure,
 }
 
 func RegisterMetrics(registry metrics.RegistererGatherer) {
@@ -117,26 +87,6 @@ func ExistingInProgress() prometheus.Counter {
 	return existingInProgress.WithLabelValues()
 }
 
-func FailureUnableToGetPod() prometheus.Counter {
-	return failureUnableToGetPod.WithLabelValues()
-}
-
-func FailurePodDoesntExist() prometheus.Counter {
-	return failurePodDoesntExist.WithLabelValues()
-}
-
-func FailureConfiguration() prometheus.Counter {
-	return failureConfiguration.WithLabelValues()
-}
-
-func FailureValidation() prometheus.Counter {
-	return failureValidation.WithLabelValues()
-}
-
-func FailureStatesDetermination() prometheus.Counter {
-	return failureStatesDetermination.WithLabelValues()
-}
-
-func FailureStatesAction() prometheus.Counter {
-	return failureStatesAction.WithLabelValues()
+func Failure(reason FailureReason) prometheus.Counter {
+	return failure.WithLabelValues(string(reason))
 }