+++ internal/metrics/reconciler/reconciler_test.go
@@ -62,56 +62,11 @@ func TestExistingInProgress(t *testing.T) {
 }
 
 func TestFailureUnableToGetPod(t *testing.T) {
-	m := FailureUnableToGetPod()
+	m := Failure("")
 	assert.Contains(
 		t,
 		m.Desc().String(),
-		fmt.Sprintf("%s_%s_%s", metricscommon.Namespace, Subsystem, failureUnableToGetPodName),
-	)
-}
-
-func TestFailurePodDoesntExist(t *testing.T) {
-	m := FailurePodDoesntExist()
-	assert.Contains(
-		t,
-		m.Desc().String(),
-		fmt.Sprintf("%s_%s_%s", metricscommon.Namespace, Subsystem, failurePodDoesntExistName),
-	)
-}
-
-func TestFailureConfiguration(t *testing.T) {
-	m := FailureConfiguration()
-	assert.Contains(
-		t,
-		m.Desc().String(),
-		fmt.Sprintf("%s_%s_%s", metricscommon.Namespace, Subsystem, failureConfigurationName),
-	)
-}
-
-func TestFailureValidation(t *testing.T) {
-	m := FailureValidation()
-	assert.Contains(
-		t,
-		m.Desc().String(),
-		fmt.Sprintf("%s_%s_%s", metricscommon.Namespace, Subsystem, failureValidationName),
-	)
-}
-
-func TestFailureStatesDetermination(t *testing.T) {
-	m := FailureStatesDetermination()
-	assert.Contains(
-		t,
-		m.Desc().String(),
-		fmt.Sprintf("%s_%s_%s", metricscommon.Namespace, Subsystem, failureStatesDeterminationName),
-	)
-}
-
-func TestFailureStatesAction(t *testing.T) {
-	m := FailureStatesAction()
-	assert.Contains(
-		t,
-		m.Desc().String(),
-		fmt.Sprintf("%s_%s_%s", metricscommon.Namespace, Subsystem, failureStatesActionName),
+		fmt.Sprintf("%s_%s_%s", metricscommon.Namespace, Subsystem, failureName),
 	)
 }